Started to implement dijkstra function. It needs some serious thinking to get it right.

+++ Graph.hs
@@ -5,7 +5,7 @@ data Graph = Graph [Elem]
 data Elem  = Elem
   { namn   :: String
   , pos    :: (Integer,Integer)
-  , adj    :: [(String, Integer)]
+  , adj    :: [(String, Double)]
   } deriving (Eq,Ord,Show)
 
 getName :: Elem -> String
@@ -14,5 +14,5 @@ getName (Elem{namn = n}) = n
 getPos :: Elem -> (Integer, Integer)
 getPos (Elem{pos = p})   = p
 
-getAdj :: Elem -> [(String, Integer)]
+getAdj :: Elem -> [(String, Double)]
 getAdj (Elem{adj = a})   = a
